6_Permanent_Contract_-_add_domain_layer: Refactor: 1. Changed Percents from decimal->float + updated extension 2. Added folders structure to Application project 3. Renamed folder Calculation -> WebApi in WebApi project 4. Shortened names of some parameters in PermanentContractSalaryCalculationContext.cs

+++ ERA.PermanentContractSalaryCalculation.Domain/Constants/PermanentContractConsts.cs
@@ -2,7 +2,7 @@
 {
     public static class PermanentContractConsts
     {
-        public const decimal CopyrightLawsCostsPercentage = 50m;
-        public const decimal TaxPercentage = 18m; // todo: dodac prog 32% itd..
+        public const float CopyrightLawsCostsPercentage = 50f;
+        public const float TaxPercentage = 18f; // todo: dodac prog 32% itd..
     }
 }
